% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryup.R
\name{get_annotations_uniprot}
\alias{get_annotations_uniprot}
\title{Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs}
\usage{
get_annotations_uniprot(id, columns = c("genes", "keywords", "families",
  "go"), updateProgress = NULL)
}
\arguments{
\item{id}{Character vector with UniProt IDs}

\item{columns}{names of uniprot data columns to retrieve. Examples include "id",
"genes", "keywords", "sequence", "go" (use \code{list_data_columns()} to see the full list)}

\item{updateProgress}{used to display progress in shiny apps}
}
\value{
a data.frame
}
\description{
Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs
}
