% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_data.R
\name{get_uniprot_data}
\alias{get_uniprot_data}
\title{Retrieve data from uniprot using uniprot's REST API}
\usage{
get_uniprot_data(
  query = NULL,
  base_url = "https://rest.uniprot.org/uniprotkb/",
  columns = c("accession", "id", "gene_names", "organism_name", "reviewed"),
  print_url = FALSE
)
}
\arguments{
\item{query}{list of keys corresponding to uniprot's query fields.
For example :
list("gene_exact" = c("Pik3r1", "Pik3r2") ,
"organism" = c("10090", "9606"), "reviewed" = "yes")}

\item{base_url}{base URL for the UniProt REST API}

\item{columns}{names of uniprot data columns to retrieve.
Examples include "accession", "id", "gene_names", "keyword", "sequence"}

\item{print_url}{logical. Prints the complete url used for the query.}
}
\value{
a data.frame
}
\description{
Retrieve data from uniprot using uniprot's REST API
}
\examples{
#Getting gene names, keywords and protein sequences for a set of UniProt IDs.
ids <- c("P22682", "P47941")
cols <- c("accession", "id", "gene_names", "keyword", "sequence")
df <- get_uniprot_data(query = list("accession_id" = ids), columns = cols)

#Lists all entries describing interactions for entry P00520.
df <- get_uniprot_data(query = list("interactor" = "P00520"), columns = cols)
}
