% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryup.R
\name{list_data_columns}
\alias{list_data_columns}
\title{list all available data columns}
\usage{
list_data_columns()
}
\description{
Prints all available data columns as
listed on the \href{https://www.uniprot.org/help/uniprotkb_column_names}{UniProt website}.
}
