% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryup.R
\name{list_query_fields}
\alias{list_query_fields}
\title{list all available query fields}
\usage{
list_query_fields()
}
\description{
Prints all available query fields as
listed on the \href{https://www.uniprot.org/help/query-fields}{UniProt website}.
}
