% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_data.R
\name{parse_messages}
\alias{parse_messages}
\title{Accessory function retrieving invalid values from messages returned by
the UniProt API.}
\usage{
parse_messages(messages)
}
\arguments{
\item{messages}{character string containing the error messages returned by
UniProt API}
}
\value{
a data.frame with invalid values (in column "value") and
corresponding query field (in column "field"). NULL if no invalid values are
identified.
}
\description{
Accessory function retrieving invalid values from messages returned by
the UniProt API.
}
