% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_fields}
\alias{query_fields}
\title{Available query fields.}
\format{
A data frame with 44 rows and 3 variables:
\describe{
  \item{field}{Name of the query field}
  \item{example}{Example query (as appearing in the query url)}
  \item{description}{Description of the example query}
}
}
\source{
\url{https://www.uniprot.org/help/query-fields}
}
\usage{
query_fields
}
\description{
Query fields that can be used to generate queries using `queryup`
along with associated examples and description.
}
\keyword{datasets}
