% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{return_fields}
\alias{return_fields}
\title{Available return fields ("columns").}
\format{
A data frame with 287 rows and 2 variables:
\describe{
  \item{field}{Name of the returned field}
  \item{label}{Label of the corresponding column in the retrieved data.frame}
}
}
\source{
\url{https://www.uniprot.org/help/return_fields}
}
\usage{
return_fields
}
\description{
Return fields that can be retrieved using `queryup`
along with their label (column "Label") as appearing in the retrieved
data.frame.
}
\keyword{datasets}
