% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryup.R
\name{get_annotations_uniprot}
\alias{get_annotations_uniprot}
\title{Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs}
\usage{
get_annotations_uniprot(id, columns = c("genes", "keywords", "families",
  "go"), max_keys = 400, updateProgress = NULL, show_progress = TRUE)
}
\arguments{
\item{id}{Character vector with UniProt IDs}

\item{columns}{names of uniprot data columns to retrieve. Examples include "id",
"genes", "keywords", "sequence", "go" (use \code{list_data_columns()} to see the full list)}

\item{max_keys}{maximum number of field items submitted}

\item{updateProgress}{used to display progress in shiny apps}

\item{show_progress}{Show progress bar}
}
\value{
a data.frame
}
\description{
Create a data.frame with UniProt annotations corrresponding to a set of UniProt IDs
}
\examples{
# Query all reviewed UniProt entries for Mus musculus:
query = list("organism" = c("10090"), "reviewed" = "yes")
df_mouse_reviewed <-  query_uniprot(query = query)
df <-  get_annotations_uniprot(id = df_mouse_reviewed$Entry[1:300], max_keys = 50)
}
