% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryup.R
\name{get_uniprot_data}
\alias{get_uniprot_data}
\title{Retrieve data from uniprot using uniprot's REST API}
\usage{
get_uniprot_data(query = NULL, columns = c("id", "genes", "organism",
  "reviewed"))
}
\arguments{
\item{query}{list of keys corresponding to uniprot's query fields. For example :
list("gene_exact" = c("Pik3r1", "Pik3r2") , "organism" = c("10090", "9606"), "reviewed" = "yes")}

\item{columns}{names of uniprot data columns to retrieve. Examples include "id",
"genes", "keywords", "sequence"}
}
\value{
a data.frame
}
\description{
Retrieve data from uniprot using uniprot's REST API
}
\examples{
#Getting gene names, keywords and protein sequences associated with a set of uniprot IDs.
ids <- c("P22682", "P47941")
cols <- c("id", "genes", "keywords", "sequence")
df <- get_uniprot_data(query = list("id" = ids), columns = cols)

#Lists all entries describing interactions with the protein described by entry P00520.
df <- get_uniprot_data(query = list("interactor" = "P00520"), columns = cols)
}
