% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryu.R
\name{query_uniprot}
\alias{query_uniprot}
\title{Retrieve data from uniprot using uniprot's REST API.
To avoid non-responsive queries, they are splitted into
smaller queries withat most \code{max_keys} items per query field.}
\usage{
query_uniprot(query = NULL, columns = c("id", "genes", "reviewed"),
  max_keys = 400)
}
\arguments{
\item{query}{list of keys corresponding to uniprot's query fields. For example :
list("gene_exact" = c("Pik3r1", "Pik3r2") , "organism" = c("10090", "9606"), "reviewed" = "yes")}

\item{columns}{names of uniprot data columns to retrieve. Examples include "id",
"genes", "keywords", "sequence"}

\item{max_keys}{maximum number of field items submitted}
}
\value{
a data.frame
}
\description{
Retrieve data from uniprot using uniprot's REST API.
To avoid non-responsive queries, they are splitted into
smaller queries withat most \code{max_keys} items per query field.
}
