% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniprot_entries}
\alias{uniprot_entries}
\title{Information for 1000 UniProt entries from the organism Mus musculus}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
  \item{Entry}{UniProt entry accession id}
  \item{Entry Name}{UniProt entry name}
  \item{Gene Names}{Gene names}
  \item{Organism (ID)}{Taxon ID}
  \item{Reviewed}{Swiss-Prot review status}
}
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
uniprot_entries
}
\description{
Entry names and other attributes of 1000 UniProt entries
in Mus musculus.
}
\keyword{datasets}
